LPS-169846 rename
