+++ workspaces/sample-minimal-workspace/client-extensions/sample-etc/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 dxp.domains=localhost:8080
-logging.level.com.easy.service=DEBUG
+logging.level.com.sample.minimal.workspace=DEBUG
 logging.level.org.springframework.security=DEBUG
 logging.level.org.springframework.web=DEBUG
-liferay.oauth.application.external.reference.code=easy-oauth-application-user-agent
+liferay.oauth.application.external.reference.code=sample-oauth-application-user-agent
 liferay.portal.url=http://localhost:8080
 server.port=58081
